fix(log): fix multiline message formatting
Styling should not cross newline boundaries,
i.e. Render(line) should never have a newline in it.
